Moving resolveSchema to new file for easier testing
